Add dist

+++ index.js
@@ -4,15 +4,17 @@ const { fs } = require('fs');
 const apiExec = __dirname.substring(0, __dirname.length - 3) + 'api\\Api.exe'
 const apiProcess = spawn(apiExec)
 
-function createWindow () 
-{
-    const win = new BrowserWindow({ width: 800, height: 600 })
+function createWindow () {
+    const win = new BrowserWindow({
+        width: 800,
+        height: 600
+    })
+
     win.setMenu(null)
     win.loadFile('index.html')
 }
 
-app.whenReady().then(() => 
-{
+app.whenReady().then(() => {
     createWindow()
 
     apiProcess.on('error', (er) => { 
@@ -20,13 +22,16 @@ app.whenReady().then(() =>
     })
 
     app.on('activate', () => {
-        if (BrowserWindow.getAllWindows().length === 0) 
+        if (BrowserWindow.getAllWindows().length === 0) {
             createWindow()
+        }
     })
-    
-    app.on('window-all-closed', () => 
+})
+
+app.on('window-all-closed', () => {
+    if (process.platform !== 'darwin')
     {
         apiProcess.kill()
         app.quit()
     }
-}
+})
